Rename: 프로젝트 컴포넌트 이름 수정
