+++ pick-front/src/project/components/ProjectSidebar.vue
